Remove old donation forms
Get rid of the now unwanted donation forms, and
remove unused code.

Ticket: https://phabricator.wikimedia.org/T421221
Bug: T421221
