--- src/components/pages/donation_form/AddressTypeBasic.vue
@@ -1,74 +1,0 @@
-<template>
-
-	<RadioField
-		name="addressType"
-		id="address-form-type"
-		:options="[
-			{
-				value: AddressTypeModel.PERSON,
-				label: $t( 'donation_form_addresstype_option_private_addresstype_basic' ),
-				id: 'addressType-0'
-			},
-			{
-				value: AddressTypeModel.COMPANY,
-				label: $t( 'donation_form_addresstype_option_company_addresstype_basic' ),
-				id: 'addressType-1'
-			},
-			{
-				value: AddressTypeModel.ANON,
-				label: $t( 'donation_form_addresstype_option_anonymous_addresstype_basic' ),
-				id: 'addressType-2'
-			},
-		]"
-		:label="$t( 'donation_form_address_choice_title_addresstype_basic' )"
-		:disabled="disabledAddressTypes"
-		:show-error="addressTypeIsInvalid"
-		:error-message="$t( 'donation_form_section_address_error' )"
-		v-model="addressType"
-		alignment="column"
-		:is-max-width-field="isMaxWidthField"
-	>
-		<template #tooltip-4>
-			<Tooltip v-if="isDirectDebit">
-				{{ $t( 'donation_form_address_choice_direct_debit_disclaimer_addresstype_basic') }}
-			</Tooltip>
-		</template>
-	</RadioField>
-
-</template>
-
-<script setup lang="ts">
-
-import RadioField from '@src/components/shared/form_fields/RadioField.vue';
-import { AddressTypeModel } from '@src/view_models/AddressTypeModel';
-import { ref, watch } from 'vue';
-import Tooltip from '@src/components/patterns/Tooltip.vue';
-
-interface Props {
-	disabledAddressTypes: AddressTypeModel[];
-	isDirectDebit: boolean;
-	addressTypeIsInvalid: boolean;
-	initialAddressType?: AddressTypeModel;
-	isMaxWidthField?: boolean;
-}
-const props = defineProps<Props>();
-
-const emit = defineEmits( [ 'address-type' ] );
-
-const addressType = ref<AddressTypeModel>( props.initialAddressType ?? AddressTypeModel.UNSET );
-
-watch( addressType, newAddressType => {
-	emit( 'address-type', newAddressType );
-} );
-
-// When disabled address type is selected, revert to person type
-watch( () => props.disabledAddressTypes, disabledAddressTypes => {
-	// TODO This assumes that person will never be disabled.
-	//      If this assumption is wrong, we need to have a fallback.
-	if ( disabledAddressTypes !== undefined &&
-			disabledAddressTypes.includes( addressType.value ) ) {
-		addressType.value = AddressTypeModel.PERSON;
-	}
-} );
-
-</script>
