--- src/components/pages/donation_form/AddressTypeIds.ts
@@ -1,9 +1,0 @@
-import { AddressTypeModel } from '@src/view_models/AddressTypeModel';
-
-export const AddressTypeIds = new Map<number, string>( [
-	[ AddressTypeModel.ANON, 'anonymous' ],
-	[ AddressTypeModel.EMAIL, 'email' ],
-	[ AddressTypeModel.PERSON, 'person' ],
-	[ AddressTypeModel.COMPANY, 'company' ],
-	[ AddressTypeModel.UNSET, 'unset' ],
-] );
