+++ tests/unit/components/pages/donation_form/DonationForm.spec.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { flushPromises, mount, VueWrapper } from '@vue/test-utils';
-import DonationForm from '@src/components/pages/donation_form/SubPages/DonationForm.vue';
+import DonationForm from '@src/components/pages/DonationForm.vue';
 import countries from '@test/data/countries';
 import type { AddressValidation } from '@src/view_models/Validation';
 import { createStore } from '@src/store/donation_store';
@@ -8,8 +8,8 @@ import { nextTick } from 'vue';
 import axios from 'axios';
 import { newSucceedingBankValidationResource } from '@test/unit/TestDoubles/SucceedingBankValidationResource';
 import { IBAN } from '@test/data/bankdata';
-import { action } from '@src/store/util';
 import { Store } from 'vuex';
+import { action } from '@src/store/util';
 import { errorSummaryItemIsFunctional } from '@test/unit/utils/errorSummaryItemIsFunctional';
 
 vi.mock( 'axios' );
@@ -27,7 +27,8 @@ declare global {
 const errorSummaryScrollElement = { scrollIntoView: () => {} };
 Object.defineProperty( document, 'getElementById', { writable: true, configurable: true, value: () => errorSummaryScrollElement } );
 
-describe( 'DonationForm.vue', () => {
+describe( 'DonationFormReceiptCompact.vue', () => {
+
 	beforeEach( () => {
 		global.window.scrollTo = vi.fn();
 		vi.useFakeTimers();
@@ -86,20 +87,12 @@ describe( 'DonationForm.vue', () => {
 		} );
 	};
 
-	it( 'sets the correct default field values', async () => {
+	it( 'initialises with the correct default field values', async () => {
 		const wrapper = getWrapper();
 
 		expect( wrapper.find<HTMLInputElement>( '#interval-0' ).element.checked ).toBeTruthy();
-		expect( wrapper.find<HTMLInputElement>( '#receipt-option-person' ).element.checked ).toBeTruthy();
-		expect( wrapper.find<HTMLInputElement>( '#person-newsletter' ).element.checked ).toBeFalsy();
-	} );
-
-	it( 'sets the correct receipt wanted field', async () => {
-		const store = createStore();
-		await store.dispatch( action( 'address', 'initializeAddress' ), { receipt: true, fields: [] } );
-		const wrapper = getWrapper( store );
-
-		expect( wrapper.find<HTMLInputElement>( '#receipt-option-person' ).element.checked ).toBeTruthy();
+		expect( wrapper.find<HTMLInputElement>( '#newsletter' ).element.checked ).toBeFalsy();
+		expect( wrapper.find<HTMLInputElement>( '#donation-receipt' ).element.checked ).toBeTruthy();
 	} );
 
 	it( 'sets the correct receipt not wanted field', async () => {
@@ -107,14 +100,14 @@ describe( 'DonationForm.vue', () => {
 		await store.dispatch( action( 'address', 'initializeAddress' ), { receipt: false, fields: [] } );
 		const wrapper = getWrapper( store );
 
-		expect( wrapper.find<HTMLInputElement>( '#receipt-option-person' ).element.checked ).toBeFalsy();
+		expect( wrapper.find<HTMLInputElement>( '#donation-receipt' ).element.checked ).toBeFalsy();
 	} );
 
-	it( 'handles the error summary when no address type was selected before submitting', async () => {
+	it( 'handles the error summary when only receipt option is yes', async () => {
 		const wrapper = getWrapper();
 
-		await wrapper.find( '#person-country' ).setValue( 'I am clearly not a country' );
-		await wrapper.find( '#person-country' ).trigger( 'blur' );
+		await wrapper.find( '#country' ).setValue( 'I am clearly not a country' );
+		await wrapper.find( '#country' ).trigger( 'blur' );
 		await vi.runAllTimersAsync();
 
 		await wrapper.find( '#submit-btn' ).trigger( 'click' );
@@ -131,82 +124,15 @@ describe( 'DonationForm.vue', () => {
 
 		expect( wrapper.find( '.error-summary' ).exists() ).toBeTruthy();
 		expect( errorSummaryItemIsFunctional( wrapper, 'amount-500', 'payment-form-amount' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'iban', 'payment-form-iban' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'addressType-0', 'address-form-type' ) ).toBeTruthy();
-
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-salutation-0', 'person-address-form-salutation' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-first-name', 'person-address-form-first-name' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-last-name', 'person-address-form-last-name' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-street', 'person-address-form-street' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-post-code', 'person-address-form-post-code' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-city', 'person-address-form-city' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-country', 'person-address-form-country' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-email', 'person-address-form-email' ) ).toBeTruthy();
-
-		await wrapper.find( 'input[name="amount"][value="500"]' ).trigger( 'change' );
-		await wrapper.find( 'input[name="addressType"][value="0"]' ).trigger( 'change' );
-		await wrapper.find( 'input[name="salutation"][value="Mr"]' ).trigger( 'change' );
-
-		await wrapper.find( '#iban' ).setValue( IBAN );
-		await wrapper.find( '#iban' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-first-name' ).setValue( 'first-name' );
-		await wrapper.find( '#person-first-name' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-last-name' ).setValue( 'last-name' );
-		await wrapper.find( '#person-last-name' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-street' ).setValue( 'street' );
-		await wrapper.find( '#person-street' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-post-code' ).setValue( '12345' );
-		await wrapper.find( '#person-post-code' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-city' ).setValue( 'city' );
-		await wrapper.find( '#person-city' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-country' ).setValue( countries[ 0 ].countryFullName );
-		await wrapper.find( '#person-country' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-email' ).setValue( 'joe@dolan.com' );
-		await wrapper.find( '#person-email' ).trigger( 'blur' );
-
-		await vi.runAllTimersAsync();
-
-		expect( wrapper.find( '.error-summary' ).exists() ).toBeFalsy();
-	} );
-
-	it( 'handles the error summary when person address type was selected before submitting', async () => {
-		const wrapper = getWrapper();
-
-		await wrapper.find( 'input[name="addressType"][value="0"]' ).trigger( 'change' );
-		await wrapper.find( '#person-country' ).setValue( 'I am clearly not a country' );
-		await wrapper.find( '#person-country' ).trigger( 'blur' );
-		await vi.runAllTimersAsync();
-
-		await wrapper.find( '#submit-btn' ).trigger( 'click' );
-		await nextTick();
-		await nextTick();
-
-		expect( errorSummaryItemIsFunctional( wrapper, 'paymentType-0', 'payment-form-type' ) ).toBeTruthy();
-
-		// Make the IBAN field appear
-		await wrapper.find( 'input[name="paymentType"][value="BEZ"]' ).trigger( 'change' );
-		await wrapper.find( '#submit-btn' ).trigger( 'click' );
-		await nextTick();
-		await nextTick();
-
-		expect( wrapper.find( '.error-summary' ).exists() ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'amount-500', 'payment-form-amount' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'iban', 'payment-form-iban' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-salutation-0', 'person-address-form-salutation' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-first-name', 'person-address-form-first-name' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-last-name', 'person-address-form-last-name' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-street', 'person-address-form-street' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-post-code', 'person-address-form-post-code' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-city', 'person-address-form-city' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-country', 'person-address-form-country' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'person-email', 'person-address-form-email' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'iban', 'payment-form-iban-calculator' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'salutation-0', 'address-form-salutation' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'first-name', 'address-form-first-name' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'last-name', 'address-form-last-name' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'email', 'address-form-email' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'street', 'address-form-street' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'post-code', 'address-form-post-code' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'city', 'address-form-city' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'country', 'address-form-country' ) ).toBeTruthy();
 
 		await wrapper.find( 'input[name="amount"][value="500"]' ).trigger( 'change' );
 		await wrapper.find( 'input[name="salutation"][value="Mr"]' ).trigger( 'change' );
@@ -214,38 +140,37 @@ describe( 'DonationForm.vue', () => {
 		await wrapper.find( '#iban' ).setValue( IBAN );
 		await wrapper.find( '#iban' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-first-name' ).setValue( 'first-name' );
-		await wrapper.find( '#person-first-name' ).trigger( 'blur' );
+		await wrapper.find( '#first-name' ).setValue( 'first-name' );
+		await wrapper.find( '#first-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-last-name' ).setValue( 'last-name' );
-		await wrapper.find( '#person-last-name' ).trigger( 'blur' );
+		await wrapper.find( '#last-name' ).setValue( 'last-name' );
+		await wrapper.find( '#last-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-street' ).setValue( 'street' );
-		await wrapper.find( '#person-street' ).trigger( 'blur' );
+		await wrapper.find( '#street' ).setValue( 'street' );
+		await wrapper.find( '#street' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-post-code' ).setValue( '12345' );
-		await wrapper.find( '#person-post-code' ).trigger( 'blur' );
+		await wrapper.find( '#post-code' ).setValue( '12345' );
+		await wrapper.find( '#post-code' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-city' ).setValue( 'city' );
-		await wrapper.find( '#person-city' ).trigger( 'blur' );
+		await wrapper.find( '#city' ).setValue( 'city' );
+		await wrapper.find( '#city' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-country' ).setValue( countries[ 0 ].countryFullName );
-		await wrapper.find( '#person-country' ).trigger( 'blur' );
+		await wrapper.find( '#country' ).setValue( countries[ 0 ].countryFullName );
+		await wrapper.find( '#country' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-email' ).setValue( 'joe@dolan.com' );
-		await wrapper.find( '#person-email' ).trigger( 'blur' );
+		await wrapper.find( '#email' ).setValue( 'joe@dolan.com' );
+		await wrapper.find( '#email' ).trigger( 'blur' );
 
 		await vi.runAllTimersAsync();
 
 		expect( wrapper.find( '.error-summary' ).exists() ).toBeFalsy();
 	} );
 
-	it( 'handles the error summary when company address type was selected before submitting', async () => {
+	it( 'handles the error summary when only receipt option is yes and company is checked', async () => {
 		const wrapper = getWrapper();
 
-		await wrapper.find( 'input[name="addressType"][value="1"]' ).trigger( 'change' );
-		await wrapper.find( '#company-country' ).setValue( 'I am clearly not a country' );
-		await wrapper.find( '#company-country' ).trigger( 'blur' );
+		await wrapper.find( '#country' ).setValue( 'I am clearly not a country' );
+		await wrapper.find( '#country' ).trigger( 'blur' );
 		await vi.runAllTimersAsync();
 
 		await wrapper.find( '#submit-btn' ).trigger( 'click' );
@@ -256,64 +181,88 @@ describe( 'DonationForm.vue', () => {
 
 		// Make the IBAN field appear
 		await wrapper.find( 'input[name="paymentType"][value="BEZ"]' ).trigger( 'change' );
+		await wrapper.find( '#is-company' ).trigger( 'click' );
 		await wrapper.find( '#submit-btn' ).trigger( 'click' );
 		await nextTick();
 		await nextTick();
 
 		expect( wrapper.find( '.error-summary' ).exists() ).toBeTruthy();
 		expect( errorSummaryItemIsFunctional( wrapper, 'amount-500', 'payment-form-amount' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'iban', 'payment-form-iban' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'company-company-name', 'company-address-form-company-name' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'company-street', 'company-address-form-street' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'company-post-code', 'company-address-form-post-code' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'company-city', 'company-address-form-city' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'company-country', 'company-address-form-country' ) ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'company-email', 'company-address-form-email' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'iban', 'payment-form-iban-calculator' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'salutation-0', 'address-form-salutation' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'first-name', 'address-form-first-name' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'last-name', 'address-form-last-name' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'email', 'address-form-email' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'street', 'address-form-street' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'company-name', 'address-form-company-name' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'post-code', 'address-form-post-code' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'city', 'address-form-city' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'country', 'address-form-country' ) ).toBeTruthy();
 
 		await wrapper.find( 'input[name="amount"][value="500"]' ).trigger( 'change' );
-		await wrapper.find( 'input[name="paymentType"][value="PPL"]' ).trigger( 'change' );
-
 		await wrapper.find( 'input[name="salutation"][value="Mr"]' ).trigger( 'change' );
 
-		await wrapper.find( '#company-company-name' ).setValue( 'company-name' );
-		await wrapper.find( '#company-company-name' ).trigger( 'blur' );
+		await wrapper.find( '#iban' ).setValue( IBAN );
+		await wrapper.find( '#iban' ).trigger( 'blur' );
 
-		await wrapper.find( '#company-street' ).setValue( 'street' );
-		await wrapper.find( '#company-street' ).trigger( 'blur' );
+		await wrapper.find( '#first-name' ).setValue( 'first-name' );
+		await wrapper.find( '#first-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#company-post-code' ).setValue( '12345' );
-		await wrapper.find( '#company-post-code' ).trigger( 'blur' );
+		await wrapper.find( '#last-name' ).setValue( 'last-name' );
+		await wrapper.find( '#last-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#company-city' ).setValue( 'city' );
-		await wrapper.find( '#company-city' ).trigger( 'blur' );
+		await wrapper.find( '#company-name' ).setValue( 'ACME' );
+		await wrapper.find( '#company-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#company-country' ).setValue( countries[ 0 ].countryFullName );
-		await wrapper.find( '#company-country' ).trigger( 'blur' );
+		await wrapper.find( '#street' ).setValue( 'street' );
+		await wrapper.find( '#street' ).trigger( 'blur' );
 
-		await wrapper.find( '#company-email' ).setValue( 'joe@dolan.com' );
-		await wrapper.find( '#company-email' ).trigger( 'blur' );
+		await wrapper.find( '#post-code' ).setValue( '12345' );
+		await wrapper.find( '#post-code' ).trigger( 'blur' );
+
+		await wrapper.find( '#city' ).setValue( 'city' );
+		await wrapper.find( '#city' ).trigger( 'blur' );
+
+		await wrapper.find( '#country' ).setValue( countries[ 0 ].countryFullName );
+		await wrapper.find( '#country' ).trigger( 'blur' );
+
+		await wrapper.find( '#email' ).setValue( 'joe@dolan.com' );
+		await wrapper.find( '#email' ).trigger( 'blur' );
 
 		await vi.runAllTimersAsync();
 
 		expect( wrapper.find( '.error-summary' ).exists() ).toBeFalsy();
 	} );
 
-	it( 'handles the error summary when without address type was selected before submitting', async () => {
+	it( 'handles the error summary when only receipt option is no', async () => {
 		const wrapper = getWrapper();
 
-		await wrapper.find( 'input[name="addressType"][value="4"]' ).trigger( 'change' );
+		await wrapper.find( '#donation-receipt' ).trigger( 'click' );
 
 		await wrapper.find( '#submit-btn' ).trigger( 'click' );
 		await nextTick();
 		await nextTick();
-		await flushPromises();
 
 		expect( wrapper.find( '.error-summary' ).exists() ).toBeTruthy();
 		expect( errorSummaryItemIsFunctional( wrapper, 'amount-500', 'payment-form-amount' ) ).toBeTruthy();
 		expect( errorSummaryItemIsFunctional( wrapper, 'paymentType-0', 'payment-form-type' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'salutation-0', 'address-form-salutation' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'first-name', 'address-form-first-name' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'last-name', 'address-form-last-name' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'email', 'address-form-email' ) ).toBeTruthy();
 
 		await wrapper.find( 'input[name="amount"][value="500"]' ).trigger( 'change' );
 		await wrapper.find( 'input[name="paymentType"][value="PPL"]' ).trigger( 'change' );
+		await wrapper.find( 'input[name="salutation"][value="Mr"]' ).trigger( 'change' );
+
+		await wrapper.find( '#first-name' ).setValue( 'first-name' );
+		await wrapper.find( '#first-name' ).trigger( 'blur' );
+
+		await wrapper.find( '#last-name' ).setValue( 'last-name' );
+		await wrapper.find( '#last-name' ).trigger( 'blur' );
+
+		await wrapper.find( '#email' ).setValue( 'joe@dolan.com' );
+		await wrapper.find( '#email' ).trigger( 'blur' );
 
 		await vi.runAllTimersAsync();
 
@@ -333,31 +282,30 @@ describe( 'DonationForm.vue', () => {
 		await wrapper.find( '#iban' ).setValue( IBAN );
 		await wrapper.find( '#iban' ).trigger( 'blur' );
 
-		await wrapper.find( 'input[name="addressType"][value="0"]' ).trigger( 'change' );
 		await wrapper.find( 'input[name="salutation"][value="Mr"]' ).trigger( 'change' );
 
-		await wrapper.find( '#person-first-name' ).setValue( 'first-name' );
-		await wrapper.find( '#person-first-name' ).trigger( 'blur' );
+		await wrapper.find( '#first-name' ).setValue( 'first-name' );
+		await wrapper.find( '#first-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-last-name' ).setValue( 'last-name' );
-		await wrapper.find( '#person-last-name' ).trigger( 'blur' );
+		await wrapper.find( '#last-name' ).setValue( 'last-name' );
+		await wrapper.find( '#last-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-street' ).setValue( 'street' );
-		await wrapper.find( '#person-street' ).trigger( 'blur' );
+		await wrapper.find( '#email' ).setValue( 'joe@dolan.com' );
+		await wrapper.find( '#email' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-post-code' ).setValue( '12345' );
-		await wrapper.find( '#person-post-code' ).trigger( 'blur' );
+		await wrapper.find( '#street' ).setValue( 'street' );
+		await wrapper.find( '#street' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-city' ).setValue( 'city' );
-		await wrapper.find( '#person-city' ).trigger( 'blur' );
+		await wrapper.find( '#post-code' ).setValue( '12345' );
+		await wrapper.find( '#post-code' ).trigger( 'blur' );
+
+		await wrapper.find( '#city' ).setValue( 'city' );
+		await wrapper.find( '#city' ).trigger( 'blur' );
 
 		await vi.runAllTimersAsync();
 
-		await wrapper.find( '#person-country' ).setValue( countries[ 0 ].countryFullName );
-		await wrapper.find( '#person-country' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-email' ).setValue( 'joe@dolan.com' );
-		await wrapper.find( '#person-email' ).trigger( 'blur' );
+		await wrapper.find( '#country' ).setValue( countries[ 0 ].countryFullName );
+		await wrapper.find( '#country' ).trigger( 'blur' );
 
 		await wrapper.find( '#submit-btn' ).trigger( 'click' );
 
@@ -380,27 +328,35 @@ describe( 'DonationForm.vue', () => {
 		await wrapper.find( '#iban' ).setValue( IBAN );
 		await wrapper.find( '#iban' ).trigger( 'blur' );
 
-		await wrapper.find( 'input[name="addressType"][value="1"]' ).trigger( 'change' );
+		await wrapper.find( 'input[name="salutation"][value="Mr"]' ).trigger( 'change' );
 
-		await wrapper.find( '#company-company-name' ).setValue( 'company-name' );
-		await wrapper.find( '#company-company-name' ).trigger( 'blur' );
+		await wrapper.find( '#first-name' ).setValue( 'first-name' );
+		await wrapper.find( '#first-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#company-street' ).setValue( 'street' );
-		await wrapper.find( '#company-street' ).trigger( 'blur' );
+		await wrapper.find( '#last-name' ).setValue( 'last-name' );
+		await wrapper.find( '#last-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#company-post-code' ).setValue( '12345' );
-		await wrapper.find( '#company-post-code' ).trigger( 'blur' );
+		await wrapper.find( '#email' ).setValue( 'joe@dolan.com' );
+		await wrapper.find( '#email' ).trigger( 'blur' );
 
-		await wrapper.find( '#company-city' ).setValue( 'city' );
-		await wrapper.find( '#company-city' ).trigger( 'blur' );
+		await wrapper.find( '#is-company' ).trigger( 'click' );
+
+		await wrapper.find( '#company-name' ).setValue( 'company-name' );
+		await wrapper.find( '#company-name' ).trigger( 'blur' );
+
+		await wrapper.find( '#street' ).setValue( 'street' );
+		await wrapper.find( '#street' ).trigger( 'blur' );
+
+		await wrapper.find( '#post-code' ).setValue( '12345' );
+		await wrapper.find( '#post-code' ).trigger( 'blur' );
+
+		await wrapper.find( '#city' ).setValue( 'city' );
+		await wrapper.find( '#city' ).trigger( 'blur' );
 
 		await vi.runAllTimersAsync();
 
-		await wrapper.find( '#company-country' ).setValue( countries[ 0 ].countryFullName );
-		await wrapper.find( '#company-country' ).trigger( 'blur' );
-
-		await wrapper.find( '#company-email' ).setValue( 'joe@dolan.com' );
-		await wrapper.find( '#company-email' ).trigger( 'blur' );
+		await wrapper.find( '#country' ).setValue( countries[ 0 ].countryFullName );
+		await wrapper.find( '#country' ).trigger( 'blur' );
 
 		await wrapper.find( '#submit-btn' ).trigger( 'click' );
 
@@ -410,7 +366,7 @@ describe( 'DonationForm.vue', () => {
 		expect( submitForm.element.submit ).toHaveBeenCalled();
 	} );
 
-	it( 'submits the form for anonymous', async () => {
+	it( 'submits the form for email-only', async () => {
 		vi.mocked( axios.post ).mockResolvedValue( { data: { status: 'OK' } } );
 		const wrapper = getWrapper();
 
@@ -418,9 +374,23 @@ describe( 'DonationForm.vue', () => {
 		submitForm.element.submit = vi.fn();
 
 		await wrapper.find( 'input[name="amount"][value="500"]' ).trigger( 'change' );
-		await wrapper.find( 'input[name="paymentType"][value="PPL"]' ).trigger( 'change' );
+		await wrapper.find( 'input[name="paymentType"][value="BEZ"]' ).trigger( 'change' );
 
-		await wrapper.find( 'input[name="addressType"][value="4"]' ).trigger( 'change' );
+		await wrapper.find( '#iban' ).setValue( IBAN );
+		await wrapper.find( '#iban' ).trigger( 'blur' );
+
+		await wrapper.find( 'input[name="salutation"][value="Mr"]' ).trigger( 'change' );
+
+		await wrapper.find( '#first-name' ).setValue( 'first-name' );
+		await wrapper.find( '#first-name' ).trigger( 'blur' );
+
+		await wrapper.find( '#last-name' ).setValue( 'last-name' );
+		await wrapper.find( '#last-name' ).trigger( 'blur' );
+
+		await wrapper.find( '#email' ).setValue( 'joe@dolan.com' );
+		await wrapper.find( '#email' ).trigger( 'blur' );
+
+		await wrapper.find( '#donation-receipt' ).trigger( 'click' );
 
 		await wrapper.find( '#submit-btn' ).trigger( 'click' );
 
@@ -440,31 +410,18 @@ describe( 'DonationForm.vue', () => {
 		await wrapper.find( 'input[name="amount"][value="500"]' ).trigger( 'change' );
 		await wrapper.find( 'input[name="paymentType"][value="UEB"]' ).trigger( 'change' );
 
-		await wrapper.find( 'input[name="addressType"][value="0"]' ).trigger( 'change' );
 		await wrapper.find( 'input[name="salutation"][value="Mr"]' ).trigger( 'change' );
 
-		await wrapper.find( '#person-first-name' ).setValue( 'first-name' );
-		await wrapper.find( '#person-first-name' ).trigger( 'blur' );
+		await wrapper.find( '#first-name' ).setValue( 'first-name' );
+		await wrapper.find( '#first-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-last-name' ).setValue( 'last-name' );
-		await wrapper.find( '#person-last-name' ).trigger( 'blur' );
+		await wrapper.find( '#last-name' ).setValue( 'last-name' );
+		await wrapper.find( '#last-name' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-street' ).setValue( 'street' );
-		await wrapper.find( '#person-street' ).trigger( 'blur' );
+		await wrapper.find( '#email' ).setValue( 'joe@dolan.com' );
+		await wrapper.find( '#email' ).trigger( 'blur' );
 
-		await wrapper.find( '#person-post-code' ).setValue( '12345' );
-		await wrapper.find( '#person-post-code' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-city' ).setValue( 'city' );
-		await wrapper.find( '#person-city' ).trigger( 'blur' );
-
-		await vi.runAllTimersAsync();
-
-		await wrapper.find( '#person-country' ).setValue( countries[ 0 ].countryFullName );
-		await wrapper.find( '#person-country' ).trigger( 'blur' );
-
-		await wrapper.find( '#person-email' ).setValue( 'joe@dolan.com' );
-		await wrapper.find( '#person-email' ).trigger( 'blur' );
+		await wrapper.find( '#donation-receipt' ).trigger( 'click' );
 
 		// change to BEZ to make the IBAN field appear
 		await wrapper.find( 'input[name="paymentType"][value="BEZ"]' ).trigger( 'change' );
@@ -475,22 +432,11 @@ describe( 'DonationForm.vue', () => {
 		await flushPromises();
 
 		expect( wrapper.find( '.error-summary' ).exists() ).toBeTruthy();
-		expect( errorSummaryItemIsFunctional( wrapper, 'iban', 'payment-form-iban' ) ).toBeTruthy();
+		expect( errorSummaryItemIsFunctional( wrapper, 'iban', 'payment-form-iban-calculator' ) ).toBeTruthy();
 
 		await wrapper.find( '#iban' ).setValue( IBAN );
 		await wrapper.find( '#iban' ).trigger( 'blur' );
 
 		expect( wrapper.find( '.error-summary' ).exists() ).toBeFalsy();
-	} );
-
-	it( 'scrolls to payment section when button for changing payment data is clicked', async () => {
-		const scrollElement = { scrollIntoView: vi.fn() };
-		Object.defineProperty( document, 'getElementById', { writable: true, configurable: true, value: () => scrollElement } );
-
-		const wrapper = getWrapper();
-
-		await wrapper.find( '#previous-btn' ).trigger( 'click' );
-
-		expect( scrollElement.scrollIntoView ).toHaveBeenCalledWith( { behavior: 'auto' } );
 	} );
 } );
