+++ tests/unit/components/pages/donation_form/Summaries/PaymentSummary.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { mount } from '@vue/test-utils';
-import PaymentSummary from '@src/components/pages/donation_form/PaymentSummary.vue';
+import PaymentSummary from '@src/components/pages/donation_form/Summaries/PaymentSummary.vue';
 
 describe( 'PaymentSummary.vue', () => {
 	it( 'renders the payment summary with paymentType', () => {
